requirements.txt
python-telegram-bot>=20.0
openai-whisper>=20230314
gspread>=5.12.0
google-auth>=2.23.0
apscheduler>=3.10.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
ffmpeg-python>=0.2.0
matplotlib>=3.8.0
pandas>=2.0.0
pytest>=7.0.0
pytz>=2023.3
